DinhNLHE140992_User_Authorization_Manager

+++ src/java/controller/ManagementAccountController.java
@@ -15,6 +15,7 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 /**
  *
@@ -35,24 +36,35 @@ public class ManagementAccountController extends HttpServlet {
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
         response.setContentType("text/html;charset=UTF-8");
-        UserDAO userDAO = new UserDAO();
-        int numberClickAccount = 1;
-        List<User> listAccount = new ArrayList<>();
-        try {
-            int userId = Integer.parseInt(request.getParameter("idDelete"));
-            if (userDAO.removeUserByUserId(userId)) {
+        HttpSession session = request.getSession();
+        User u = (User) session.getAttribute("user");
+        if (u != null) {
+            if (u.getRoleId() == 1) {
+                UserDAO userDAO = new UserDAO();
+            int numberClickAccount = 1;
+            List<User> listAccount = new ArrayList<>();
+            try {
+                int userId = Integer.parseInt(request.getParameter("idDelete"));
+                if (userDAO.removeUserByUserId(userId)) {
+                    listAccount = userDAO.getAllUsers();
+                    //Set data to JSP
+                    request.setAttribute("numberClickAccount", numberClickAccount);
+                    request.setAttribute("LIST_User", listAccount);
+                    request.getRequestDispatcher("dashboard/dashboardAccount.jsp").forward(request, response);
+                }
+            } catch (Exception e) {
                 listAccount = userDAO.getAllUsers();
                 //Set data to JSP
                 request.setAttribute("numberClickAccount", numberClickAccount);
                 request.setAttribute("LIST_User", listAccount);
                 request.getRequestDispatcher("dashboard/dashboardAccount.jsp").forward(request, response);
             }
-        } catch (Exception e) {
-            listAccount = userDAO.getAllUsers();
-            //Set data to JSP
-            request.setAttribute("numberClickAccount", numberClickAccount);
-            request.setAttribute("LIST_User", listAccount);
-            request.getRequestDispatcher("dashboard/dashboardAccount.jsp").forward(request, response);
+            }
+            else {
+                request.getRequestDispatcher("error.jsp").forward(request, response);
+            }
+        } else {
+            request.getRequestDispatcher("error.jsp").forward(request, response);
         }
     }
 
@@ -68,17 +80,27 @@ public class ManagementAccountController extends HttpServlet {
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
         int numberClickAccount = 1;
-        try {
-            String username = ("".equals(request.getParameter("username"))) ? "" : request.getParameter("username");
-            String email = ("".equals(request.getParameter("email"))) ? "" : request.getParameter("email");
-            String password = ("".equals(request.getParameter("password"))) ? "" : request.getParameter("password");
-            int roleId = Integer.parseInt(("".equals(request.getParameter("roleId"))) ? "" : request.getParameter("roleId"));
-            UserDAO uDAO = new UserDAO();
-            if (uDAO.createUser(username, password, email, roleId, 1)) {
-                request.setAttribute("numberClickAccount", numberClickAccount);
-                response.sendRedirect("dashboard-account");
+        HttpSession session = request.getSession();
+        User u = (User) session.getAttribute("user");
+        if (u != null) {
+            if (u.getRoleId() == 1) {
+                try {
+                    String username = ("".equals(request.getParameter("username"))) ? "" : request.getParameter("username");
+                    String email = ("".equals(request.getParameter("email"))) ? "" : request.getParameter("email");
+                    String password = ("".equals(request.getParameter("password"))) ? "" : request.getParameter("password");
+                    int roleId = Integer.parseInt(("".equals(request.getParameter("roleId"))) ? "" : request.getParameter("roleId"));
+                    UserDAO uDAO = new UserDAO();
+                    if (uDAO.createUser(username, password, email, roleId, 1)) {
+                        request.setAttribute("numberClickAccount", numberClickAccount);
+                        response.sendRedirect("dashboard-account");
+                    }
+                } catch (Exception e) {
+                    request.getRequestDispatcher("error.jsp").forward(request, response);
+                }
+            } else {
+                request.getRequestDispatcher("error.jsp").forward(request, response);
             }
-        } catch (Exception e) {
+        } else {
             request.getRequestDispatcher("error.jsp").forward(request, response);
         }
     }
