+++ src/java/controller/ManagementProductController.java
@@ -54,14 +54,40 @@ public class ManagementProductController extends HttpServlet {
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
         int numberClickProduct = 1;
-        String id = request.getParameter("id");
-        ProductDAO pDAO = new ProductDAO();
-        
-        request.setAttribute("stt", null);
+        HttpSession session = request.getSession();
+        User u = (User) session.getAttribute("user");
+        if (u != null) {
+            if (u.getRoleId() == 1 || u.getRoleId() == 2) {
+                String id = request.getParameter("id");
+                ProductDAO pDAO = new ProductDAO();
 
-        if (id != null) {
-            int idDel = Integer.parseInt(id);
-            if (pDAO.deleteProductByProductId(idDel)) {
+                request.setAttribute("stt", null);
+
+                if (id != null) {
+                    int idDel = Integer.parseInt(id);
+                    if (pDAO.deleteProductByProductId(idDel)) {
+                        List<Product> listProduct = new ArrayList<>();
+                        listProduct = pDAO.getAllProductInformation();
+                        request.setAttribute("listProduct", listProduct);
+
+                        CategoryDAO cDAO = new CategoryDAO();
+                        List<Category> listCategory = new ArrayList<>();
+                        listCategory = cDAO.getAll();
+                        request.setAttribute("listCategory", listCategory);
+                        request.setAttribute("numberClickProduct", numberClickProduct);
+                        request.setAttribute("stt", "3");
+                        request.getRequestDispatcher("dashboard/dashboardProduct.jsp").forward(request, response);
+                    } else {
+                        CategoryDAO cDAO = new CategoryDAO();
+                        List<Category> listCategory = new ArrayList<>();
+                        listCategory = cDAO.getAll();
+                        request.setAttribute("listCategory", listCategory);
+                        request.setAttribute("numberClickProduct", numberClickProduct);
+                        request.setAttribute("stt", "4");
+                        request.getRequestDispatcher("dashboard/dashboardProduct.jsp").forward(request, response);
+                    }
+                }
+
                 List<Product> listProduct = new ArrayList<>();
                 listProduct = pDAO.getAllProductInformation();
                 request.setAttribute("listProduct", listProduct);
@@ -71,29 +97,13 @@ public class ManagementProductController extends HttpServlet {
                 listCategory = cDAO.getAll();
                 request.setAttribute("listCategory", listCategory);
                 request.setAttribute("numberClickProduct", numberClickProduct);
-                request.setAttribute("stt", "3");
                 request.getRequestDispatcher("dashboard/dashboardProduct.jsp").forward(request, response);
             } else {
-                CategoryDAO cDAO = new CategoryDAO();
-                List<Category> listCategory = new ArrayList<>();
-                listCategory = cDAO.getAll();
-                request.setAttribute("listCategory", listCategory);
-                request.setAttribute("numberClickProduct", numberClickProduct);
-                request.setAttribute("stt", "4");
-                request.getRequestDispatcher("dashboard/dashboardProduct.jsp").forward(request, response);
+                request.getRequestDispatcher("error.jsp").forward(request, response);
             }
+        } else {
+            request.getRequestDispatcher("error.jsp").forward(request, response);
         }
-
-        List<Product> listProduct = new ArrayList<>();
-        listProduct = pDAO.getAllProductInformation();
-        request.setAttribute("listProduct", listProduct);
-
-        CategoryDAO cDAO = new CategoryDAO();
-        List<Category> listCategory = new ArrayList<>();
-        listCategory = cDAO.getAll();
-        request.setAttribute("listCategory", listCategory);
-        request.setAttribute("numberClickProduct", numberClickProduct);
-        request.getRequestDispatcher("dashboard/dashboardProduct.jsp").forward(request, response);
     }
 
     /**
@@ -107,92 +117,100 @@ public class ManagementProductController extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        String name = request.getParameter("name");
-        Double price = Double.parseDouble(("".equals(request.getParameter("price")) ? "0" : request.getParameter("price")));
-        int categoryId = Integer.parseInt(("".equals(request.getParameter("category")) ? "0" : request.getParameter("category")));
-        String imgSqu = request.getParameter("imageSqu");
-        String imgRec = request.getParameter("imageRec");
-        String desc = request.getParameter("description");
-        List<String> listProductKey = new ArrayList<>();
-        int numberClickProduct = 1;
-        //Up file to Upload folder
-        Part uploadFilePart = request.getPart("product-key-file");
-        String uploadFileName = Paths.get(uploadFilePart.getSubmittedFileName()).getFileName().toString();
-
-        InputStream inputStream = uploadFilePart.getInputStream();
-
-        String uploadFolderPath = getServletContext().getRealPath("/upload") + File.separator + uploadFileName;
-
-        try {
-            byte[] byt = new byte[inputStream.available()];
-            inputStream.read(byt);
-            FileOutputStream fileOutputStream = new FileOutputStream(uploadFolderPath);
-            fileOutputStream.write(byt);
-            fileOutputStream.flush();
-            fileOutputStream.close();
-
-            //Read data from Uploaded file
-            FileInputStream fileInputStream = new FileInputStream(uploadFolderPath);
-            XSSFWorkbook workbook = new XSSFWorkbook(fileInputStream);
-            XSSFSheet sheet = workbook.getSheetAt(0);
-            FormulaEvaluator formula = workbook.getCreationHelper().createFormulaEvaluator();
-            for (Row row : sheet) {
-                for (Cell cell : row) {
-                    if (!cell.getStringCellValue().equals("")) {
-                        listProductKey.add(cell.getStringCellValue());
-                    }
-                }
-            }
-            workbook.close();
-            fileInputStream.close();
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-
-        //Add new Product
-        ProductDAO pDAO = new ProductDAO();
-
         HttpSession session = request.getSession();
         User u = (User) session.getAttribute("user");
+        if (u != null) {
+            if (u.getRoleId() == 1 || u.getRoleId() == 2) {
+                String name = request.getParameter("name");
+                Double price = Double.parseDouble(("".equals(request.getParameter("price")) ? "0" : request.getParameter("price")));
+                int categoryId = Integer.parseInt(("".equals(request.getParameter("category")) ? "0" : request.getParameter("category")));
+                String imgSqu = request.getParameter("imageSqu");
+                String imgRec = request.getParameter("imageRec");
+                String desc = request.getParameter("description");
+                List<String> listProductKey = new ArrayList<>();
+                int numberClickProduct = 1;
+                //Up file to Upload folder
+                Part uploadFilePart = request.getPart("product-key-file");
+                String uploadFileName = Paths.get(uploadFilePart.getSubmittedFileName()).getFileName().toString();
 
-        if (pDAO.createNewProduct(name, desc, price, categoryId, u.getUserId(), listProductKey.size())) {
-            int newestId = pDAO.getNewestProductId();
-            if (pDAO.insertProductImageSquare(newestId, imgSqu) && pDAO.insertProductImageRectangle(newestId, imgRec)) {
-                for (String key : listProductKey) {
-                    //Encrypt product key before insert to database
-                    String encryptKey = AES.encrypt(key, "@SWP391_Group2");
-                    pDAO.insertProductKey(newestId, encryptKey);
+                InputStream inputStream = uploadFilePart.getInputStream();
+
+                String uploadFolderPath = getServletContext().getRealPath("/upload") + File.separator + uploadFileName;
+
+                try {
+                    byte[] byt = new byte[inputStream.available()];
+                    inputStream.read(byt);
+                    FileOutputStream fileOutputStream = new FileOutputStream(uploadFolderPath);
+                    fileOutputStream.write(byt);
+                    fileOutputStream.flush();
+                    fileOutputStream.close();
+
+                    //Read data from Uploaded file
+                    FileInputStream fileInputStream = new FileInputStream(uploadFolderPath);
+                    XSSFWorkbook workbook = new XSSFWorkbook(fileInputStream);
+                    XSSFSheet sheet = workbook.getSheetAt(0);
+                    FormulaEvaluator formula = workbook.getCreationHelper().createFormulaEvaluator();
+                    for (Row row : sheet) {
+                        for (Cell cell : row) {
+                            if (!cell.getStringCellValue().equals("")) {
+                                listProductKey.add(cell.getStringCellValue());
+                            }
+                        }
+                    }
+                    workbook.close();
+                    fileInputStream.close();
+                } catch (Exception e) {
+                    e.printStackTrace();
                 }
-                request.setAttribute("stt", "1");
-            } else {
+
+                //Add new Product
+                ProductDAO pDAO = new ProductDAO();
+
+                if (pDAO.createNewProduct(name, desc, price, categoryId, u.getUserId(), listProductKey.size())) {
+                    int newestId = pDAO.getNewestProductId();
+                    if (pDAO.insertProductImageSquare(newestId, imgSqu) && pDAO.insertProductImageRectangle(newestId, imgRec)) {
+                        for (String key : listProductKey) {
+                            //Encrypt product key before insert to database
+                            String encryptKey = AES.encrypt(key, "@SWP391_Group2");
+                            pDAO.insertProductKey(newestId, encryptKey);
+                        }
+                        request.setAttribute("stt", "1");
+                    } else {
+                        CategoryDAO cDAO = new CategoryDAO();
+                        List<Category> listCategory = new ArrayList<>();
+                        listCategory = cDAO.getAll();
+                        request.setAttribute("listCategory", listCategory);
+                        request.setAttribute("numberClickProduct", numberClickProduct);
+                        request.setAttribute("stt", "2");
+                        request.getRequestDispatcher("dashboard/dashboardProduct.jsp").forward(request, response);
+                    }
+                } else {
+                    CategoryDAO cDAO = new CategoryDAO();
+                    List<Category> listCategory = new ArrayList<>();
+                    listCategory = cDAO.getAll();
+                    request.setAttribute("listCategory", listCategory);
+                    request.setAttribute("numberClickProduct", numberClickProduct);
+                    request.setAttribute("stt", "2");
+                    request.getRequestDispatcher("dashboard/dashboardProduct.jsp").forward(request, response);
+                }
+
+                List<Product> listProduct = new ArrayList<>();
+                listProduct = pDAO.getAllProductInformation();
+                request.setAttribute("listProduct", listProduct);
+
                 CategoryDAO cDAO = new CategoryDAO();
                 List<Category> listCategory = new ArrayList<>();
                 listCategory = cDAO.getAll();
                 request.setAttribute("listCategory", listCategory);
                 request.setAttribute("numberClickProduct", numberClickProduct);
-                request.setAttribute("stt", "2");
                 request.getRequestDispatcher("dashboard/dashboardProduct.jsp").forward(request, response);
+            } else {
+                request.getRequestDispatcher("error.jsp").forward(request, response);
             }
         } else {
-            CategoryDAO cDAO = new CategoryDAO();
-            List<Category> listCategory = new ArrayList<>();
-            listCategory = cDAO.getAll();
-            request.setAttribute("listCategory", listCategory);
-            request.setAttribute("numberClickProduct", numberClickProduct);
-            request.setAttribute("stt", "2");
-            request.getRequestDispatcher("dashboard/dashboardProduct.jsp").forward(request, response);
+            request.getRequestDispatcher("error.jsp").forward(request, response);
         }
 
-        List<Product> listProduct = new ArrayList<>();
-        listProduct = pDAO.getAllProductInformation();
-        request.setAttribute("listProduct", listProduct);
-
-        CategoryDAO cDAO = new CategoryDAO();
-        List<Category> listCategory = new ArrayList<>();
-        listCategory = cDAO.getAll();
-        request.setAttribute("listCategory", listCategory);
-        request.setAttribute("numberClickProduct", numberClickProduct);
-        request.getRequestDispatcher("dashboard/dashboardProduct.jsp").forward(request, response);
     }
 
     /**
